requirements.txt
asgiref==3.4.1
cffi==1.14.6
cryptography==3.4.7
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==3.2.5
django-cache-url==3.2.3
django-environ==0.4.5
django-filter==21.1
django-rest-knox==4.1.0
djangorestframework==3.12.4
environs==9.3.2
marshmallow==3.12.2
psycopg2==2.9.1
pycparser==2.20
python-dotenv==0.18.0
pytz==2021.1
sqlparse==0.4.1
